Adding expectedReturn and Annualized Standard Deviation for the Tickers

+++ src/main/java/com/ms/portfoliomanager/processor/GBMotionCalculator.java
@@ -4,11 +4,11 @@ import org.apache.commons.math3.distribution.NormalDistribution;
 
 public class GBMotionCalculator {
 
-    public static double geometricMotion(double mu, double sigma,int time,double price,double nd){
+    public static double geometricMotion(double mu, double sigma, double time, double price, double nd){
         NormalDistribution normalDistribution = new NormalDistribution(mu ,
                 sigma);
         nd = normalDistribution.inverseCumulativeProbability(nd);
-        double part = (double) time/7257600;
+        double part = time/7257600;
         double muPart = mu*part;
         double ndPart = (sigma*nd)*Math.sqrt(part);
         double delS = price*(muPart+ndPart);
